# perf: Capture InstanceTypeFilterErrors in error structs and don't print them until the end