<!-- Please follow the guidelines at https://www.conventionalcommits.org/en/v1.0.0/ and use one of the following in your title:
feat:            <-- New features that require a MINOR version update
fix:             <-- Bug fixes that require at PATCH version update
chore:           <-- Smaller changes that impact behavior but aren't large enough to be features
perf:            <-- Code changes that improve performance but do not impact behavior
docs:            <-- Documentation changes that do not impact code
test:            <-- Test changes that do not impact behavior
ci:              <-- Changes that affect test or rollout automation
!${type}:        <-- Include ! if your change includes a backwards incompatible change.
-->

Fixes #N/A <!-- issue number -->

**Description**

This change ensures that we don't perform expensive operations that show-up in the CPU profile and show-up in our heap analysis to print the nodeClaimRequirements and performing resources.Merge. 

These changes were executed against the scheduling benchmark (which showed negligible performance changes) and a live test that deploys 20,000 pods to the cluster and generates 10,000 nodes (with two pods per node constrained by the size of the instance types on NodePools). Prior to the change, this test took __31m49s__. After the change, the same test case takes __26m34s__ (a 5m15s improvement).

![flame_requirements_improvement](https://github.com/user-attachments/assets/3b1b230d-7b0f-44f0-bd28-e622b7f872b4)

### Before PR

#### Scheduling Benchmark

```
=== RUN   TestSchedulingProfile
scheduled 40151 against 8031 nodes in total in 59.198867831s 678.2393223232317 pods/sec
400 instances 1 pods      1 nodes     183.666µs per scheduling      183.666µs per pod
400 instances 50 pods     10 nodes    42.342407ms per scheduling    846.848µs per pod
400 instances 100 pods    20 nodes    86.044652ms per scheduling    860.446µs per pod
400 instances 500 pods    100 nodes   408.015194ms per scheduling   816.03µs per pod
400 instances 1000 pods   200 nodes   803.296687ms per scheduling   803.296µs per pod
400 instances 1500 pods   300 nodes   1.23015975s per scheduling    820.106µs per pod
400 instances 2000 pods   400 nodes   1.744999875s per scheduling   872.499µs per pod
400 instances 5000 pods   1000 nodes  5.008729875s per scheduling   1.001745ms per pod
400 instances 10000 pods  2000 nodes  12.685723333s per scheduling  1.268572ms per pod
400 instances 20000 pods  4000 nodes  35.421737833s per scheduling  1.771086ms per pod
--- PASS: TestSchedulingProfile (67.02s)
PASS
```

#### Live Test 

__Scheduling Time: 31m49s__

```
{"level":"INFO","time":"2025-02-02T06:45:17.502Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":3563,"pods-remaining":16437,"duration":"1m0s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T06:46:17.563Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":5017,"pods-remaining":14983,"duration":"2m0s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T06:47:17.663Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":6127,"pods-remaining":13873,"duration":"3m0s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T06:48:17.718Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":7071,"pods-remaining":12929,"duration":"4m0s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T06:49:17.873Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":7911,"pods-remaining":12089,"duration":"5m0s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T06:50:18.015Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":8671,"pods-remaining":11329,"duration":"6m0s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T06:51:18.075Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":9367,"pods-remaining":10633,"duration":"7m0s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T06:52:18.319Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":10017,"pods-remaining":9983,"duration":"8m0s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T06:53:18.497Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":10627,"pods-remaining":9373,"duration":"9m1s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T06:54:18.551Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":11205,"pods-remaining":8795,"duration":"10m1s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T06:55:18.745Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":11755,"pods-remaining":8245,"duration":"11m1s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T06:56:18.965Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":12283,"pods-remaining":7717,"duration":"12m1s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T06:57:19.157Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":12787,"pods-remaining":7213,"duration":"13m1s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T06:58:19.252Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":13273,"pods-remaining":6727,"duration":"14m1s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T06:59:19.280Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":13743,"pods-remaining":6257,"duration":"15m1s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:00:19.325Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":14195,"pods-remaining":5805,"duration":"16m1s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:01:19.376Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":14635,"pods-remaining":5365,"duration":"17m1s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:02:19.579Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":15063,"pods-remaining":4937,"duration":"18m2s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:03:19.664Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":15479,"pods-remaining":4521,"duration":"19m2s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:04:19.813Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":15885,"pods-remaining":4115,"duration":"20m2s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:05:20.038Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":16281,"pods-remaining":3719,"duration":"21m2s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:06:20.256Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":16663,"pods-remaining":3337,"duration":"22m2s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:07:20.293Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":17035,"pods-remaining":2965,"duration":"23m2s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:08:20.542Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":17407,"pods-remaining":2593,"duration":"24m3s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:09:20.691Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":17763,"pods-remaining":2237,"duration":"25m3s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:10:20.868Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":18117,"pods-remaining":1883,"duration":"26m3s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:11:20.904Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":18463,"pods-remaining":1537,"duration":"27m3s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:12:21.082Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":18795,"pods-remaining":1205,"duration":"28m3s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:13:21.410Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":19123,"pods-remaining":877,"duration":"29m3s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:14:21.578Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":19447,"pods-remaining":553,"duration":"30m4s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:15:21.892Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","pods-scheduled":19767,"pods-remaining":233,"duration":"31m4s","scheduling-id":"cbdf5dcc-94c3-4371-b0a6-f8574137b2af"}
{"level":"INFO","time":"2025-02-02T07:16:06.064Z","logger":"controller","caller":"provisioning/provisioner.go:129","message":"found provisionable pod(s)","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","Pods":"default/test-51, default/test-123, default/test-31, default/test-181, default/test-19 and 19995 other(s)","duration":"31m49.389166009s"}
{"level":"INFO","time":"2025-02-02T07:16:06.094Z","logger":"controller","caller":"provisioning/provisioner.go:350","message":"computed new nodeclaim(s) to fit pod(s)","commit":"16dfcb0-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"c3bd7c1f-4074-4d75-83e3-22c046e7efe6","nodeclaims":10000,"pods":20000}
```

#### Heap Profile

![requirements_improvement](https://github.com/user-attachments/assets/edaa5460-5218-4c9b-98c7-83aa546b6eea)

### After PR

#### Scheduling Benchmark

```
=== RUN   TestSchedulingProfile
scheduled 40151 against 8031 nodes in total in 59.329537061s 676.7455468044276 pods/sec
400 instances 1 pods      1 nodes     199.111µs per scheduling      199.111µs per pod
400 instances 50 pods     10 nodes    42.671329ms per scheduling    853.426µs per pod
400 instances 100 pods    20 nodes    110.710599ms per scheduling   1.107105ms per pod
400 instances 500 pods    100 nodes   432.48425ms per scheduling    864.968µs per pod
400 instances 1000 pods   200 nodes   1.021048083s per scheduling   1.021048ms per pod
400 instances 1500 pods   300 nodes   1.41967825s per scheduling    946.452µs per pod
400 instances 2000 pods   400 nodes   1.755113042s per scheduling   877.556µs per pod
400 instances 5000 pods   1000 nodes  5.021272375s per scheduling   1.004254ms per pod
400 instances 10000 pods  2000 nodes  12.540354458s per scheduling  1.254035ms per pod
400 instances 20000 pods  4000 nodes  35.597483625s per scheduling  1.779874ms per pod
--- PASS: TestSchedulingProfile (66.63s)
PASS
```

#### Live Test

__Scheduling Time: 26m34s__

```
{"level":"INFO","time":"2025-02-02T06:14:35.499Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"0edfada-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"aea398a8-2f52-4a1d-9250-70b4a268231e","pods-scheduled":3993,"pods-remaining":16007,"duration":"1m0s","scheduling-id":"7045ca91-106e-453a-8921-3cd9b2719ba1"}
{"level":"INFO","time":"2025-02-02T06:15:35.613Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"0edfada-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"aea398a8-2f52-4a1d-9250-70b4a268231e","pods-scheduled":5631,"pods-remaining":14369,"duration":"2m0s","scheduling-id":"7045ca91-106e-453a-8921-3cd9b2719ba1"}
{"level":"INFO","time":"2025-02-02T06:16:35.706Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"0edfada-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"aea398a8-2f52-4a1d-9250-70b4a268231e","pods-scheduled":6837,"pods-remaining":13163,"duration":"3m0s","scheduling-id":"7045ca91-106e-453a-8921-3cd9b2719ba1"}
{"level":"INFO","time":"2025-02-02T06:17:35.835Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"0edfada-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"aea398a8-2f52-4a1d-9250-70b4a268231e","pods-scheduled":7859,"pods-remaining":12141,"duration":"4m0s","scheduling-id":"7045ca91-106e-453a-8921-3cd9b2719ba1"}
{"level":"INFO","time":"2025-02-02T06:18:35.920Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"0edfada-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"aea398a8-2f52-4a1d-9250-70b4a268231e","pods-scheduled":8763,"pods-remaining":11237,"duration":"5m0s","scheduling-id":"7045ca91-106e-453a-8921-3cd9b2719ba1"}
{"level":"INFO","time":"2025-02-02T06:19:36.030Z","logger":"controller","caller":"provisioning/provisioner.go:343","message":"computing pod scheduling...","commit":"0edfada-dirty","controller":"provisioner","namespace":"","name":"","reconcileID":"aea398a8-2f52-4a1d-9250-70b4a268231e","pods-scheduled":9581,"pods-remaining":10419,"duration":"6m0s","scheduling-id":"7045ca91-106e-453a-8921-3cd9b2719ba1"}
…